完成用户管理和材料分类管理接口

+++ src/main/java/com/wang/material/material/entity/UserEntity.java
@@ -21,7 +21,7 @@ public class UserEntity extends BaseEntity{
     private String password;
 
     @ApiModelProperty("电话")
-    private Long phone;
+    private Integer phone;
 
     @ApiModelProperty("地址")
     private String address;
